Fix "warning: cast from pointer to integer of different size".
git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@38463 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ gui/dialog/dialog.c
@@ -327,7 +327,7 @@ void gtkShow(int type, char *param)
             gtk_widget_destroy(PopUpMenu);
         }
 
-        PopUpMenu = CreatePopUpMenu((int)param);
+        PopUpMenu = CreatePopUpMenu((intptr_t)param);
         gtk_menu_popup(GTK_MENU(PopUpMenu), NULL, NULL, NULL, NULL, 0, 0);
         break;
